build-release: Support repos with multiple distribution packages
For repos with multiple distribution packages, they're still released
individually by running `build-release` in the directory with the
package's pyproject.toml. But if there's no `pactivate` in the current
working directory, we look for `../pactivate` and use the shared
../.build/ and its virtualenv.

This removes the previous support for giving a subdirectory name,
which superficially had a similar effect but didn't actually deal with
multiple packages in one .build/release/ dir properly. We might add
this back at a later point if we feel it's useful not to have to `cd`.

Amongst other changes, this now requires that our logfile names
include the name of the distribution package and that when we clean
before building we remove only the logfile and package files for this
package, not all packages.